+++ packages/server-v4/src/routes/v4/page/addInitScript.ts
@@ -1,0 +1,35 @@
+import type { RouteOptions } from "fastify";
+import { Api } from "@browserbasehq/stagehand";
+import type { FastifyZodOpenApiSchema } from "fastify-zod-openapi";
+
+import {
+  PageAddInitScriptActionSchema,
+  PageAddInitScriptRequestSchema,
+  PageAddInitScriptResponseSchema,
+} from "../../../schemas/v4/page.js";
+import { createPageActionHandler, pageErrorResponses } from "./shared.js";
+
+const addInitScriptRoute: RouteOptions = {
+  method: "POST",
+  url: "/page/addInitScript",
+  schema: {
+    operationId: "PageAddInitScript",
+    summary: "page.addInitScript",
+    headers: Api.SessionHeadersSchema,
+    body: PageAddInitScriptRequestSchema,
+    response: {
+      200: PageAddInitScriptResponseSchema,
+      ...pageErrorResponses,
+    },
+  } satisfies FastifyZodOpenApiSchema,
+  handler: createPageActionHandler({
+    method: "addInitScript",
+    actionSchema: PageAddInitScriptActionSchema,
+    execute: async ({ page, params }) => {
+      await page.addInitScript(params.script);
+      return { added: true };
+    },
+  }),
+};
+
+export default addInitScriptRoute;
