+++ packages/server-v4/src/routes/v4/page/close.ts
@@ -1,0 +1,35 @@
+import type { RouteOptions } from "fastify";
+import { Api } from "@browserbasehq/stagehand";
+import type { FastifyZodOpenApiSchema } from "fastify-zod-openapi";
+
+import {
+  PageCloseActionSchema,
+  PageCloseRequestSchema,
+  PageCloseResponseSchema,
+} from "../../../schemas/v4/page.js";
+import { createPageActionHandler, pageErrorResponses } from "./shared.js";
+
+const closeRoute: RouteOptions = {
+  method: "POST",
+  url: "/page/close",
+  schema: {
+    operationId: "PageClose",
+    summary: "page.close",
+    headers: Api.SessionHeadersSchema,
+    body: PageCloseRequestSchema,
+    response: {
+      200: PageCloseResponseSchema,
+      ...pageErrorResponses,
+    },
+  } satisfies FastifyZodOpenApiSchema,
+  handler: createPageActionHandler({
+    method: "close",
+    actionSchema: PageCloseActionSchema,
+    execute: async ({ page }) => {
+      await page.close();
+      return { closed: true };
+    },
+  }),
+};
+
+export default closeRoute;
