+++ packages/server-v4/src/routes/v4/page/frames.ts
@@ -1,0 +1,41 @@
+import type { RouteOptions } from "fastify";
+import { Api } from "@browserbasehq/stagehand";
+import type { FastifyZodOpenApiSchema } from "fastify-zod-openapi";
+
+import {
+  PageFramesActionSchema,
+  PageFramesRequestSchema,
+  PageFramesResponseSchema,
+} from "../../../schemas/v4/page.js";
+import { createPageActionHandler, pageErrorResponses } from "./shared.js";
+
+const framesRoute: RouteOptions = {
+  method: "POST",
+  url: "/page/frames",
+  schema: {
+    operationId: "PageFrames",
+    summary: "page.frames",
+    headers: Api.SessionHeadersSchema,
+    body: PageFramesRequestSchema,
+    response: {
+      200: PageFramesResponseSchema,
+      ...pageErrorResponses,
+    },
+  } satisfies FastifyZodOpenApiSchema,
+  handler: createPageActionHandler({
+    method: "frames",
+    actionSchema: PageFramesActionSchema,
+    execute: async ({ page }) => {
+      return {
+        frames: page.frames().map((frame) => ({
+          frameId: frame.frameId,
+          pageId: frame.pageId,
+          sessionId: frame.sessionId,
+          isBrowserRemote: frame.isBrowserRemote(),
+        })),
+      };
+    },
+  }),
+};
+
+export default framesRoute;
