+++ packages/server-v4/src/routes/v4/page/listAllFrameIds.ts
@@ -1,0 +1,34 @@
+import type { RouteOptions } from "fastify";
+import { Api } from "@browserbasehq/stagehand";
+import type { FastifyZodOpenApiSchema } from "fastify-zod-openapi";
+
+import {
+  PageListAllFrameIdsActionSchema,
+  PageListAllFrameIdsRequestSchema,
+  PageListAllFrameIdsResponseSchema,
+} from "../../../schemas/v4/page.js";
+import { createPageActionHandler, pageErrorResponses } from "./shared.js";
+
+const listAllFrameIdsRoute: RouteOptions = {
+  method: "POST",
+  url: "/page/listAllFrameIds",
+  schema: {
+    operationId: "PageListAllFrameIds",
+    summary: "page.listAllFrameIds",
+    headers: Api.SessionHeadersSchema,
+    body: PageListAllFrameIdsRequestSchema,
+    response: {
+      200: PageListAllFrameIdsResponseSchema,
+      ...pageErrorResponses,
+    },
+  } satisfies FastifyZodOpenApiSchema,
+  handler: createPageActionHandler({
+    method: "listAllFrameIds",
+    actionSchema: PageListAllFrameIdsActionSchema,
+    execute: async ({ page }) => {
+      return { frameIds: page.listAllFrameIds() };
+    },
+  }),
+};
+
+export default listAllFrameIdsRoute;
