+++ packages/server-v4/src/routes/v4/page/url.ts
@@ -1,0 +1,34 @@
+import type { RouteOptions } from "fastify";
+import { Api } from "@browserbasehq/stagehand";
+import type { FastifyZodOpenApiSchema } from "fastify-zod-openapi";
+
+import {
+  PageUrlActionSchema,
+  PageUrlRequestSchema,
+  PageUrlResponseSchema,
+} from "../../../schemas/v4/page.js";
+import { createPageActionHandler, pageErrorResponses } from "./shared.js";
+
+const urlRoute: RouteOptions = {
+  method: "POST",
+  url: "/page/url",
+  schema: {
+    operationId: "PageUrl",
+    summary: "page.url",
+    headers: Api.SessionHeadersSchema,
+    body: PageUrlRequestSchema,
+    response: {
+      200: PageUrlResponseSchema,
+      ...pageErrorResponses,
+    },
+  } satisfies FastifyZodOpenApiSchema,
+  handler: createPageActionHandler({
+    method: "url",
+    actionSchema: PageUrlActionSchema,
+    execute: async ({ page }) => {
+      return { url: page.url() };
+    },
+  }),
+};
+
+export default urlRoute;
